+++ cmd/project.go
@@ -142,6 +142,10 @@ func (p *ProjectCommands) createProjectFile() error {
 
 	p.appendScopes()
 
+	if p.Ctx.IsSet("sops-age-keys") {
+		p.projectFile.Spec.ProjectSopsAgeKeys = p.Ctx.StringSlice("sops-age-keys")
+	}
+
 	client, err := github.NewClient(clusterDepsRepoOwner, clusterDepsRepo, "", github.APIBaseURL)
 	if err != nil {
 		return err