+++ cmd/secret_generation.go
@@ -3,15 +3,12 @@ package cmd
 import (
 	"bytes"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
 	"strings"
-	"sync"
 	"text/template"
 
 	"github.com/Masterminds/sprig/v3"
-	"github.com/helmfile/vals"
 	"go.uber.org/zap"
 	"golang.org/x/crypto/ssh/terminal"
 	"gopkg.in/yaml.v3"
@@ -27,8 +24,6 @@ const (
 	FetchSecretValue = "fetchSecretValue"
 	Prompt           = "prompt"
 	RequiredEnv      = "requiredEnv"
-
-	valsCacheSize = 256
 )
 
 type GenerationSpec struct {
@@ -46,9 +41,6 @@ type GenerationRule struct {
 	Name     string `yaml:"name"`
 	Template string `yaml:"template"`
 }
-
-var instance *vals.Runtime
-var once sync.Once
 
 func prompt(name string) (string, error) {
 	fmt.Printf("Enter %s: ", name)
@@ -69,58 +61,10 @@ func requiredEnv(name string) (string, error) {
 	return "", fmt.Errorf("required env var %s is not set", name)
 }
 
-func valsInstance() (*vals.Runtime, error) {
-	var err error
-	once.Do(func() {
-		instance, err = vals.New(vals.Options{CacheSize: valsCacheSize, LogOutput: io.Discard})
-	})
-
-	return instance, err
-}
-
-func fetchSecretValue(path string) (string, error) {
-	valsMap := make(map[string]any)
-	valsMap["key"] = path
-	resultMap, err := expandSecretRefs(valsMap)
-	if err != nil {
-		return "", err
-	}
-
-	rendered, ok := resultMap["key"]
-	if !ok {
-		return "", fmt.Errorf("unexpected error occurred, %v doesn't have 'key' key", resultMap)
-	}
-
-	result, ok := rendered.(string)
-	if !ok {
-		return "", fmt.Errorf("expected %v to be string", rendered)
-	}
-
-	return result, nil
-}
-
-func expandSecretRefs(values map[string]any) (map[string]any, error) {
-	awsEnvs := map[string]string{
-		aws_provider.AWSSDKLoadConfig: "1",
-		aws_provider.AWSSDKGoLogLevel: "off",
-	}
-
-	if err := util.SetOSEnvs(false, awsEnvs); err != nil {
-		return nil, err
-	}
-
-	runtime, err := valsInstance()
-	if err != nil {
-		return nil, err
-	}
-
-	return runtime.Eval(values)
-}
-
 func (gf *GenerationFuncMap) createFuncMap() {
 	gf.funcMap = sprig.TxtFuncMap()
 	for key, val := range map[string]interface{}{
-		FetchSecretValue: fetchSecretValue,
+		FetchSecretValue: util.ValsFetchSecretValue,
 		Prompt:           prompt,
 		RequiredEnv:      requiredEnv,
 	} {