+++ docs/release-notes.md
@@ -2,3 +2,4 @@
 - Extended release category to support Helmfile v1.1.X command set.
 - Added fetchSecretValue function for using Vals in secret spec templates.
 - Added cluster provider credentials environment variables for runtime Helmfile.
+- Added project.yaml option spec.sops-age-keys to fetch SOPS age keys via Vals backends.